add support for detecting default config
The idea is that an internal package could contain the
simple config implementation so all of the internal uses
would automatically get the right behavior.

Also adds support for automatically starting the global
registry and stopping on exit. That can be useful for
libraries that may use spectator with a default config
and cannot rely on the app owner using the with block
properly.